src/requirements.txt
Flask==2.3.2
Flask-Session2
Flask-WTF
gunicorn
pymongo==3.9.0
Jinja2
gitpython==3.1.29
